+++ kitty/boss.py
@@ -213,7 +213,7 @@ class Boss:
                 os_window_id = create_os_window(
                         initial_window_size_func(size_data, self.cached_values),
                         pre_show_callback,
-                        self.args.title or appname, wname, wclass)
+                        self.args.title or appname, wname, wclass, disallow_override_title=bool(self.args.title))
         else:
             wname = self.args.name or self.args.cls or appname
             wclass = self.args.cls or appname